Done Notification and tested, added Privacy Policy

+++ lib/models/models.dart
@@ -277,7 +277,7 @@ class Driver {
       dateOfBirth: DateTime.fromMillisecondsSinceEpoch(map['dateOfBirth'] ?? 0),
       gender: Utils.getGenderFromString(map['gender']),
       rank: map['rank'],
-      rating: map['rating'] as num ?? 0,
+      rating: (map['rating'] as num).toDouble() ?? 0,
       totalReviews: map['totalReviews'] ?? 0,
       balance: map['balance'] != null
           ? (map['balance'].runtimeType == int
